+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/table/TestTableComponents.scala
@@ -31,7 +31,7 @@ object TestTableComponents {
       oneRecordTableSourceName,
       new TableSourceFactory(
         tableDefs(oneRecordTableName),
-        enableFlinkBatchExecutionMode = false,
+        enableFlinkBatchExecutionMode = true,
         testDataGenerationMode = TestDataGenerationMode.Live
       )
     )