+++ docs/Changelog.md
@@ -43,6 +43,7 @@
     * Added ability to run tests on data from file or generated on the spot (the `generated` button)
   * [#6518](https://github.com/TouK/nussknacker/pull/6518) Added user friendly editor for Table Sink output value
     * The editor can be easily switched backed to 'raw' version similarly to the current Kafka Sink
+  * [#6545](https://github.com/TouK/nussknacker/pull/6545) Virtual columns are properly handled in Table Sink
 
 1.16.2 (18 July 2024)
 -------------------------